+++ cmd/operator-sdk/generate/csv.go
@@ -16,16 +16,13 @@ package generate
 
 import (
 	"fmt"
-	"io/ioutil"
 	"path/filepath"
 
 	"github.com/operator-framework/operator-sdk/internal/generate/gen"
 	gencatalog "github.com/operator-framework/operator-sdk/internal/generate/olm-catalog"
-	"github.com/operator-framework/operator-sdk/internal/scaffold"
-	"github.com/operator-framework/operator-sdk/internal/util/k8sutil"
 	"github.com/operator-framework/operator-sdk/internal/util/projutil"
 
-	"github.com/coreos/go-semver/semver"
+	"github.com/blang/semver"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -41,8 +38,10 @@ type csvCmd struct {
 	crdDir         string
 	updateCRDs     bool
 	defaultChannel bool
+	makeManifests  bool
 }
 
+//nolint:lll
 func newGenerateCSVCmd() *cobra.Command {
 	c := &csvCmd{}
 	cmd := &cobra.Command{
@@ -55,19 +54,26 @@ A CSV semantic version is supplied via the --csv-version flag. If your operator
 has already generated a CSV manifest you want to use as a base, supply its
 version to --from-version. Otherwise the SDK will scaffold a new CSV manifest.
 
-CSV input flags:
-	--deploy-dir:
-		The CSV's install strategy and permissions will be generated from the operator manifests
-		(Deployment and Role/ClusterRole) present in this directory.
+The --make-manifests flag directs the generator to create a 'manifests' directory
+intended to hold your latest operator manifests. This flag is true by default.
 
-	--apis-dir:
-		The CSV annotation comments will be parsed from the Go types under this path to
-		fill out metadata for owned APIs in spec.customresourcedefinitions.owned.
+More information on manifests:
+https://github.com/operator-framework/operator-registry/blob/master/docs/design/operator-bundle.md#operator-bundle-overview
 
-	--crd-dir:
-		The CSV's spec.customresourcedefinitions.owned field is generated from the CRD manifests
-		in this path.These CRD manifests are also copied over to the bundle directory if --update-crds is set.
-		Additionally the CR manifests will be used to populate the CSV example CRs.
+Flags that change project default paths:
+  --deploy-dir:
+    The CSV's install strategy and permissions will be generated from the operator manifests
+    (Deployment and Role/ClusterRole) present in this directory.
+
+  --apis-dir:
+    The CSV annotation comments will be parsed from the Go types under this path to
+    fill out metadata for owned APIs in spec.customresourcedefinitions.owned.
+
+  --crd-dir:
+    The CSV's spec.customresourcedefinitions.owned field is generated from the CRD manifests
+    in this path. These CRD manifests are also copied over to the bundle directory if
+    --update-crds=true (the default). Additionally the CR manifests will be used to populate
+    the CSV example CRs.
 `,
 		Example: `
 		##### Generate CSV from default input paths #####
@@ -149,6 +155,11 @@ CSV input flags:
 					"runs from the project root directory.")
 			}
 
+			// Legacy behavior.
+			if !c.makeManifests && !cmd.Flags().Changed("update-crds") {
+				c.updateCRDs = false
+			}
+
 			if err := c.run(); err != nil {
 				log.Fatal(err)
 			}
@@ -157,12 +168,14 @@ CSV input flags:
 	}
 
 	cmd.Flags().StringVar(&c.csvVersion, "csv-version", "",
-		"Semantic version of the CSV")
-	if err := cmd.MarkFlagRequired("csv-version"); err != nil {
-		log.Fatalf("Failed to mark `csv-version` flag for `generate csv` subcommand as required: %v", err)
-	}
+		"Semantic version of the CSV. This flag must be set if a package manifest exists")
 	cmd.Flags().StringVar(&c.fromVersion, "from-version", "",
 		"Semantic version of an existing CSV to use as a base")
+	err := cmd.Flags().MarkDeprecated("from-version",
+		"Use --csv-version to update your bundled CSV in `manifests/`")
+	if err != nil {
+		panic(err)
+	}
 
 	// TODO: Allow multiple paths
 	// Deployment and RBAC manifests might be in different dirs e.g kubebuilder
@@ -175,18 +188,37 @@ CSV input flags:
 	cmd.Flags().StringVar(&c.crdDir, "crd-dir", "",
 		`Project relative path to root directory for CRD and CR manifests`)
 
-	cmd.Flags().StringVar(&c.outputDir, "output-dir", scaffold.DeployDir,
-		"Base directory to output generated CSV. The resulting CSV bundle directory "+
-			"will be \"<output-dir>/olm-catalog/<operator-name>/<csv-version>\".")
-	cmd.Flags().BoolVar(&c.updateCRDs, "update-crds", false,
-		"Update CRD manifests in deploy/{operator-name}/{csv-version} the using latest API's")
+	cmd.Flags().StringVar(&c.outputDir, "output-dir", "",
+		"Base directory to output generated CSV. If --make-manifests=false the resulting "+
+			"CSV bundle directory will be <output-dir>/olm-catalog/<operator-name>/<csv-version>. "+
+			"If --make-manifests=true, the bundle directory will be <output-dir>/manifests")
 	cmd.Flags().StringVar(&c.operatorName, "operator-name", "",
 		"Operator name to use while generating CSV")
 	cmd.Flags().StringVar(&c.csvChannel, "csv-channel", "",
 		"Channel the CSV should be registered under in the package manifest")
+	err = cmd.Flags().MarkDeprecated("csv-channel", "Package manifests are deprecated. "+
+		"Run `operator-sdk bundle create --generate-only` to create operator metadata")
+	if err != nil {
+		panic(err)
+	}
 	cmd.Flags().BoolVar(&c.defaultChannel, "default-channel", false,
 		"Use the channel passed to --csv-channel as the package manifests' default channel. "+
 			"Only valid when --csv-channel is set")
+	err = cmd.Flags().MarkDeprecated("default-channel", "Package manifests are deprecated. "+
+		"Run `operator-sdk bundle create --generate-only` to create operator metadata")
+	if err != nil {
+		panic(err)
+	}
+
+	cmd.Flags().BoolVar(&c.updateCRDs, "update-crds", true,
+		"Update CRD manifests in deploy/<operator-name>/<csv-version> from the default "+
+			"CRDs dir deploy/crds or --crd-dir if set. If --make-manifests=false, this option "+
+			"is false by default")
+	cmd.Flags().BoolVar(&c.makeManifests, "make-manifests", true,
+		"When set, the generator will create or update a CSV manifest in a 'manifests' "+
+			"directory. This directory is intended to be used for your latest bundle manifests. "+
+			"The default location is deploy/olm-catalog/<operator-name>/manifests. "+
+			"If --output-dir is set, the directory will be <output-dir>/manifests")
 
 	return cmd
 }
@@ -210,7 +242,7 @@ func (c csvCmd) run() error {
 	cfg := gen.Config{
 		OperatorName: c.operatorName,
 		// TODO(hasbro17): Remove the Input key map when the Generator input keys
-		// are removed in favour of config fields in the csvGenerator
+		// are removed in favour of config fields in the bundle generator
 		Inputs: map[string]string{
 			gencatalog.DeployDirKey: c.deployDir,
 			gencatalog.APIsDirKey:   c.apisDir,
@@ -219,7 +251,7 @@ func (c csvCmd) run() error {
 		OutputDir: c.outputDir,
 	}
 
-	csv := gencatalog.NewCSV(cfg, c.csvVersion, c.fromVersion)
+	csv := gencatalog.NewBundle(cfg, c.csvVersion, c.fromVersion, c.updateCRDs, c.makeManifests)
 	if err := csv.Generate(); err != nil {
 		return fmt.Errorf("error generating CSV: %v", err)
 	}
@@ -228,30 +260,27 @@ func (c csvCmd) run() error {
 		return fmt.Errorf("error generating package manifest: %v", err)
 	}
 
-	// Write CRD's to the new or updated CSV package dir.
-	if c.updateCRDs {
-		// TODO: This path should come from the CSV generator field csvOutputDir
-		bundleDir := filepath.Join(c.outputDir, gencatalog.OLMCatalogChildDir, c.operatorName, c.csvVersion)
-		if err := copyCustomResourceDefinitions(c.crdDir, bundleDir); err != nil {
-			return err
-		}
-	}
+	log.Info("CSV manifest generated successfully")
 
 	return nil
 }
 
 func (c csvCmd) validate() error {
-	if err := validateVersion(c.csvVersion); err != nil {
-		return err
+	// If a manifests directory exists, allow no versions to be set. In this case
+	// either a new CSV will be created or existing CSV updated in the manifests dir.
+	if c.csvVersion != "" {
+		if err := validateVersion(c.csvVersion); err != nil {
+			return err
+		}
 	}
 	if c.fromVersion != "" {
 		if err := validateVersion(c.fromVersion); err != nil {
 			return err
 		}
-	}
-	if c.fromVersion != "" && c.csvVersion == c.fromVersion {
-		return fmt.Errorf(
-			"from-version (%s) cannot equal csv-version; set only csv-version instead", c.fromVersion)
+		if c.csvVersion == c.fromVersion {
+			return fmt.Errorf("--from-version (%s) cannot equal --csv-version; set only csv-version instead",
+				c.fromVersion)
+		}
 	}
 
 	if c.defaultChannel && c.csvChannel == "" {
@@ -262,7 +291,7 @@ func (c csvCmd) validate() error {
 }
 
 func validateVersion(version string) error {
-	v, err := semver.NewVersion(version)
+	v, err := semver.Parse(version)
 	if err != nil {
 		return fmt.Errorf("%s is not a valid semantic version: %v", version, err)
 	}
@@ -271,52 +300,5 @@ func validateVersion(version string) error {
 	if v.String() != version {
 		return fmt.Errorf("provided CSV version %s contains bad values (parses to %s)", version, v)
 	}
-	return nil
-}
-
-// copyCustomResourceDefinitions copies all CustomResourceDefinition manifests
-// from fromDir to toDir with the same file name.
-func copyCustomResourceDefinitions(fromDir, toDir string) error {
-	infos, err := ioutil.ReadDir(fromDir)
-	if err != nil {
-		return err
-	}
-
-	for _, info := range infos {
-		if info.IsDir() {
-			continue
-		}
-
-		fromPath := filepath.Join(fromDir, info.Name())
-		b, err := ioutil.ReadFile(fromPath)
-		if err != nil {
-			return fmt.Errorf("error reading manifest %s: %v", fromPath, err)
-		}
-
-		scanner := k8sutil.NewYAMLScanner(b)
-		manifests := []byte{}
-		for scanner.Scan() {
-			manifest := scanner.Bytes()
-			typeMeta, err := k8sutil.GetTypeMetaFromBytes(manifest)
-			if err != nil {
-				log.Debugf("Skipping non-manifest file %s: %v", fromPath, err)
-				continue
-			}
-			if typeMeta.Kind == "CustomResourceDefinition" {
-				manifests = k8sutil.CombineManifests(manifests, b)
-			}
-		}
-		if err = scanner.Err(); err != nil {
-			return err
-		}
-
-		if len(manifests) != 0 {
-			toPath := filepath.Join(toDir, info.Name())
-			if err = ioutil.WriteFile(toPath, manifests, info.Mode()); err != nil {
-				return fmt.Errorf("error writing CRD %s: %v", toPath, err)
-			}
-		}
-	}
-
 	return nil
 }